Refactored views
Refactored views
