+++ app/src/main/AndroidManifest.xml
@@ -12,9 +12,9 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".SecurityQuestionsActivity"></activity>
-        <activity android:name=".ProvideBankAccountNumberActivity" />
-        <activity android:name=".SelectYourBankActivity" />
+        <activity android:name=".views.ui.security_questions.SecurityQuestionsActivity"></activity>
+        <activity android:name=".views.ui.provide_bank_account_number.ProvideBankAccountNumberActivity" />
+        <activity android:name=".views.ui.select_your_bank.SelectYourBankActivity" />
         <activity
             android:name=".views.MainActivity"
             android:label="@string/app_name"
