--- app/src/main/java/com/planetpay/SelectYourBankActivity.java
@@ -1,14 +1,0 @@
-package com.planetpay;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class SelectYourBankActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_select_your_bank);
-    }
-}
