+++ app/src/main/java/com/planetpay/views/ui/security_questions/SecurityQuestionsActivity.java
@@ -1,0 +1,16 @@
+package com.planetpay.views.ui.security_questions;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.planetpay.R;
+
+public class SecurityQuestionsActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_security_questions);
+    }
+}
