+++ app/src/main/java/com/planetpay/views/ui/select_your_bank/SelectYourBankActivity.java
@@ -1,14 +1,16 @@
-package com.planetpay;
+package com.planetpay.views.ui.select_your_bank;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
-public class SecurityQuestionsActivity extends AppCompatActivity {
+import com.planetpay.R;
+
+public class SelectYourBankActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_security_questions);
+        setContentView(R.layout.activity_select_your_bank);
     }
 }
