+++ app/src/main/res/layout/activity_security_questions.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SecurityQuestionsActivity">
+    tools:context=".views.ui.security_questions.SecurityQuestionsActivity">
 
     <ScrollView
         android:layout_width="match_parent"
